fix(routes): rename
