+++ app/modules/utils/routes/normalizeRouteSearchText.ts
