+++ app/pages/Routes.test.tsx
@@ -8,7 +8,7 @@ import { AreaType } from '~/modules/enums/AreaType'
 import { AppLocaleType } from '~/modules/enums/AppLocaleType'
 import { CityNameType } from '~/modules/enums/CityNameType'
 import routeSearchSlice from '~/modules/slices/routeSearchSlice'
-import { ROUTE_SEARCH_RECENT_STORAGE_KEY } from '~/modules/utils/routeSearch/routeSearchRecentStorage'
+import { ROUTE_SEARCH_RECENT_STORAGE_KEY } from '~/modules/utils/routes/routeSearchRecentStorage'
 import { createTestStore } from '~/test/createTestStore'
 import { renderWithProvidersAndRouter } from '~/test/render'
 import Routes from './Routes'
